+++ Minimal/Code/Debug/DebugFunctions.py
@@ -26,10 +26,10 @@ def set_world_size(size: _float) -> None:
 type SimulateUnlocks = _dict[Unlock, _int] | _tuple[_tuple[Unlock, _int]] | _list[_tuple[Unlock, _int]] | _tuple[Unlock] | _list[Unlock] | Unlocks
 
 def simulate(
-		filename: _str,
+		filename: string,
 		sim_unlocks: SimulateUnlocks,
 		sim_items: _dict[Item, _float],
-		sim_globals: _dict[_str, Any],
+		sim_globals: _dict[string, Any],
 		seed: _float, speedup: _float
 	) -> _float:
 	# Docstring: simulate