+++ Minimal/Code/Enums/Grounds.py
@@ -6,8 +6,12 @@ class Ground:
 
 # --------------------------------------------------
 class Grounds:
-	Grassland: Ground
+	@staticmethod
+	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Ground:
+		return Ground()
+
+	Grassland = _auto()
 	# Docstring: Grassland
 
-	Soil: Ground
+	Soil = _auto()
 	# Docstring: Soil