+++ Minimal/Code/Enums/Hats.py
@@ -6,80 +6,84 @@ class Hat:
 
 # --------------------------------------------------
 class Hats:
-	Brown_Hat: Hat
+	@staticmethod
+	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Hat:
+		return Hat()
+
+	Brown_Hat = _auto()
 	# Docstring: Brown_Hat
 
-	Cactus_Hat: Hat
+	Cactus_Hat = _auto()
 	# Docstring: Cactus_Hat
 
-	Carrot_Hat: Hat
+	Carrot_Hat = _auto()
 	# Docstring: Carrot_Hat
 
-	Dinosaur_Hat: Hat
+	Dinosaur_Hat = _auto()
 	# Docstring: Dinosaur_Hat
 
-	Gold_Hat: Hat
+	Gold_Hat = _auto()
 	# Docstring: Gold_Hat
 
-	Gold_Trophy_Hat: Hat
+	Gold_Trophy_Hat = _auto()
 	# Docstring: Gold_Trophy_Hat
 
-	Golden_Cactus_Hat: Hat
+	Golden_Cactus_Hat = _auto()
 	# Docstring: Golden_Cactus_Hat
 
-	Golden_Carrot_Hat: Hat
+	Golden_Carrot_Hat = _auto()
 	# Docstring: Golden_Carrot_Hat
 
-	Golden_Gold_Hat: Hat
+	Golden_Gold_Hat = _auto()
 	# Docstring: Golden_Gold_Hat
 
-	Golden_Pumpkin_Hat: Hat
+	Golden_Pumpkin_Hat = _auto()
 	# Docstring: Golden_Pumpkin_Hat
 
-	Golden_Sunflower_Hat: Hat
+	Golden_Sunflower_Hat = _auto()
 	# Docstring: Golden_Sunflower_Hat
 
-	Golden_Tree_Hat: Hat
+	Golden_Tree_Hat = _auto()
 	# Docstring: Golden_Tree_Hat
 
-	Gray_Hat: Hat
+	Gray_Hat = _auto()
 	# Docstring: Gray_Hat
 
-	Green_Hat: Hat
+	Green_Hat = _auto()
 	# Docstring: Green_Hat
 
-	Pumpkin_Hat: Hat
+	Pumpkin_Hat = _auto()
 	# Docstring: Pumpkin_Hat
 
-	Purple_Hat: Hat
+	Purple_Hat = _auto()
 	# Docstring: Purple_Hat
 
-	Silver_Trophy_Hat: Hat
+	Silver_Trophy_Hat = _auto()
 	# Docstring: Silver_Trophy_Hat
 
-	Straw_Hat: Hat
+	Straw_Hat = _auto()
 	# Docstring: Straw_Hat
 
-	Sunflower_Hat: Hat
+	Sunflower_Hat = _auto()
 	# Docstring: Sunflower_Hat
 
-	The_Farmers_Remains: Hat
+	The_Farmers_Remains = _auto()
 	# Docstring: The_Farmers_Remains
 
-	Top_Hat: Hat
+	Top_Hat = _auto()
 	# Docstring: Top_Hat
 
-	Traffic_Cone: Hat
+	Traffic_Cone = _auto()
 	# Docstring: Traffic_Cone
 
-	Traffic_Cone_Stack: Hat
+	Traffic_Cone_Stack = _auto()
 	# Docstring: Traffic_Cone_Stack
 
-	Tree_Hat: Hat
+	Tree_Hat = _auto()
 	# Docstring: Tree_Hat
 
-	Wizard_Hat: Hat
+	Wizard_Hat = _auto()
 	# Docstring: Wizard_Hat
 
-	Wood_Trophy_Hat: Hat
+	Wood_Trophy_Hat = _auto()
 	# Docstring: Wood_Trophy_Hat