+++ Minimal/Code/Enums/Leaderboard.py
@@ -6,50 +6,54 @@ class Leaderboard:
 
 # --------------------------------------------------
 class Leaderboards:
-	Cactus: Leaderboard
+	@staticmethod
+	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Leaderboard:
+		return Leaderboard()
+
+	Cactus = _auto()
 	# Docstring: Cactus (Leaderboard)
 
-	Cactus_Single: Leaderboard
+	Cactus_Single = _auto()
 	# Docstring: Cactus_Single
 
-	Carrots: Leaderboard
+	Carrots = _auto()
 	# Docstring: Carrots (Leaderboard)
 
-	Carrots_Single: Leaderboard
+	Carrots_Single = _auto()
 	# Docstring: Carrots_Single
 
-	Dinosaur: Leaderboard
+	Dinosaur = _auto()
 	# Docstring: Dinosaur (Leaderboard)
 
-	Fastest_Reset: Leaderboard
+	Fastest_Reset = _auto()
 	# Docstring: Fastest_Reset
 
-	Hay: Leaderboard
+	Hay = _auto()
 	# Docstring: Hay (Leaderboard)
 
-	Hay_Single: Leaderboard
+	Hay_Single = _auto()
 	# Docstring: Hay_Single
 
-	Maze: Leaderboard
+	Maze = _auto()
 	# Docstring: Maze (Leaderboard)
 
-	Maze_Single: Leaderboard
+	Maze_Single = _auto()
 	# Docstring: Maze_Single
 
-	Pumpkins: Leaderboard
+	Pumpkins = _auto()
 	# Docstring: Pumpkins (Leaderboard)
 
-	Pumpkins_Single: Leaderboard
+	Pumpkins_Single = _auto()
 	# Docstring: Pumpkins_Single
 
-	Sunflowers: Leaderboard
+	Sunflowers = _auto()
 	# Docstring: Sunflowers (Leaderboard)
 
-	Sunflowers_Single: Leaderboard
+	Sunflowers_Single = _auto()
 	# Docstring: Sunflowers_Single
 
-	Wood: Leaderboard
+	Wood = _auto()
 	# Docstring: Wood (Leaderboard)
 
-	Wood_Single: Leaderboard
+	Wood_Single = _auto()
 	# Docstring: Wood_Single