+++ Tests/MinimalTest.py
@@ -1,0 +1,3 @@
+from Release.__builtins__minimal import *
+
+Unlocks.