+++ Tests/TypingRecognition_Minimal.py
@@ -46,7 +46,13 @@ TestSet = set()
 
 dict.len(TestList) # Error
 dict.len(TestDict) # Proper usage
+pop(dict({"1":1}), 1)
 
+
+
+TestListInt : list[int] = list()
+TestListInt.append(1)
+list(TestListInt)
 
 list(1) # Error - wrong type
 list(1, 2) # Error - positional
@@ -69,9 +75,9 @@ append([1], 2)
 list({1})
 list(list([1]))
 
-from builtins import (bool as _bool, int as _int, float as _float, str as _str,
-					  range as _range,
-					  tuple as _tuple, list as _list, set as _set, dict as _dict)
+# from builtins import (bool as _bool, int as _int, float as _float, str as string,
+# 					  range as range_class,
+# 					  tuple as _tuple, list as _list, set as _set, dict as _dict)
 
 from typing import Sequence
 
@@ -117,9 +123,49 @@ ItemDict[Items.Power] = 1000
 
 simulate("Hi", UnlockDict, ItemDict, {}, 1, 250)
 
-MyList: list[int] = list()
+MyList: list[int] = list([1])
+
+ListAny : list[Any] = [1, "hhih", 1.1]
 
 remove(MyList, 1)
 insert(MyList, 1, 1)
 
+MySet : set[int] = {1, 2, 3}
+
+remove(MySet, 1)
+
 print(list([1]))
+print(((1, 2), (3, 3)))
+print(({1, 2}, (3, 3)))
+print({(1, 2), (3, 4)})
+
+TestString : string = str(1)
+
+class Custom():
+      ...
+
+TestSetCustom : set[Custom]
+
+simulate()
+
+Literal[1]
+
+TestAny : Any = ((1, 2, 3), 1, 2.3, "string", ((1, 2), (3, "Hey")), True, False)
+TestAny1 : Any = 1
+Test1Hashable : Hashable = 1
+TestAnyHashable : Hashable = (((0, 1), 2, 3), (1, 2, 3))
+TestDictHashable : dict[Hashable, Any] = {(1, 2, 3): 1}
+
+Primitive
+Enums
+Hashable
+Any
+AnyIterable
+
+
+TestDictAnyHuge : dict[Hashable, Any] = {1: [1, 2, 3, (1, 2, 3), False, None, {1, 2, 3}]}
+TestDictSmall : dict[Hashable, Any] = {1: 1}
+
+Entities.Apple
+for entity in Entities:
+	print(entity)